Remove PID parameters; integrate into control parameters.
We dropped the uniform parameter structs in favour of a single big parameter struct with all the variables, with different possible meanings (maximums or integral gains). We also drop the pid calculation in favour of simple integral controllers because the cycle frequency is high enough that we always want to average out the noise over a few cycles. Thus the proportional and derivative terms are too noisy for meaningful control, but at 23khz the integral only control converges very quickly so we don't need the extra terms. This lets us free up some compute and reduce a piece of mutable state.